Implement custom IntoReponse error type
Currently it implements From<anyhow::Error> and From<http::StatusCode>,
which means we can use ? to propagate errors easily inside handlers.

Note that it specifically hides the error message from the user, which
is good security practice just-in-case. Future work may be needed on
intercepting errors from extrators to hide the messages as well.