Fix add_to_calendar route
This commit fixes some issues caused by concurrent work on the store and
add_to_calendar route. It now uses courses instead of sections, and
takes in the term as part of the URL. At also changes the HTMX target
to the courses container, which should set up the next UI work.