Add basic selected courses view
I also wired up the database to the views, so it should "just work".

The database stuff is getting messy so I want to refactor soon, I'm
thinking that we need to separate the database type from the database
connection type, hopefully to reduce pulling from the pool in every
method.